Revert "add halfway not done yet"
This reverts commit 89df247a4ad7aea55f6fdc0a2fe9a528648ed9fa.
