--- include/common.h
@@ -1,6 +1,0 @@
-// User files
-#include "main.h"
-
-// Options file
-#include "options.h"
-#include "chassis.h"
